refactor: move shared config components from `shared/components/qbd/` to `shared/components/configuration/`
