+++ src/app/shared/components/configuration/configuration-label/configuration-label.component.html
