+++ src/app/shared/components/configuration/configuration-radio-field/configuration-radio-field.component.ts
